update for cython3 and python3.12

+++ gomea/real_valued.pyx
@@ -24,17 +24,17 @@ cdef class RealValuedGOMEA:
         fitness: FitnessFunction, 
         # GOMEA parameters
         linkage_model : LinkageModel = StaticLinkageTree(),
-        lower_init_range: double=0.0,
-        upper_init_range: double=1.0,
+        lower_init_range: float = 0.0,
+        upper_init_range: float = 1.0,
         random_seed: int=-1,
         # IMS settings (optional)
-        max_number_of_populations: int=25,
-        max_number_of_generations: int=-1,
-        IMS_subgeneration_factor: int=8,
-        base_population_size: int=10,
+        max_number_of_populations: int = 25,
+        max_number_of_generations: int = -1,
+        IMS_subgeneration_factor: int = 8,
+        base_population_size: int = 10,
         # Termination settings (optional)
-        max_number_of_evaluations : double=-1,
-        max_number_of_seconds : double=-1,
+        max_number_of_evaluations : float = -1,
+        max_number_of_seconds : float = -1,
     ):
         
         # Initialize attributes
